+++ src/main/java/com/kafkapingpong/infrastructure/repository/PongProducerRepository.java
@@ -2,20 +2,22 @@ package com.kafkapingpong.infrastructure.repository;
 
 import com.kafkapingpong.domain.message.Message;
 import com.kafkapingpong.domain.message.PongRepository;
-import com.kafkapingpong.infrastructure.producer.ErrorProducer;
-import com.kafkapingpong.infrastructure.producer.SuccessProducer;
+import com.kafkapingpong.infrastructure.producer.MessageOutProducer;
+import com.kafkapingpong.infrastructure.repository.dto.MessageOut;
+import com.kafkapingpong.infrastructure.repository.dto.PayloadError;
+import com.kafkapingpong.infrastructure.repository.dto.PayloadSuccess;
 
 import java.time.Duration;
 
 public class PongProducerRepository implements PongRepository {
-  private final SuccessProducer pongSuccessProducer;
-  private final ErrorProducer pongErrorProducer;
-  private final ErrorProducer dlqProducer;
+  private final MessageOutProducer pongSuccessProducer;
+  private final MessageOutProducer pongErrorProducer;
+  private final MessageOutProducer dlqProducer;
 
   public PongProducerRepository(
-      SuccessProducer pongSuccessProducer,
-      ErrorProducer pongErrorProducer,
-      ErrorProducer dlqProducer) {
+      MessageOutProducer pongSuccessProducer,
+      MessageOutProducer pongErrorProducer,
+      MessageOutProducer dlqProducer) {
     this.pongSuccessProducer = pongSuccessProducer;
     this.pongErrorProducer = pongErrorProducer;
     this.dlqProducer = dlqProducer;
@@ -23,16 +25,28 @@ public class PongProducerRepository implements PongRepository {
 
   @Override
   public void pong(Message message, Duration duration) {
-    pongSuccessProducer.sendSuccess(message, duration);
+    pongSuccessProducer.send(mapToPongSuccess(message, duration));
   }
 
   @Override
   public void pongForError(Message message) {
-    pongErrorProducer.sendError(message);
+    pongErrorProducer.send(mapToPongError(message));
   }
 
   @Override
   public void dlq(Message message) {
-    dlqProducer.sendError(message);
+    dlqProducer.send(mapToPongError(message));
+  }
+
+  private MessageOut mapToPongSuccess(Message message, Duration duration) {
+    return new MessageOut(
+        message.getTransactionId().toString(),
+        new PayloadSuccess("pong", duration.toMillis()));
+  }
+
+  private MessageOut mapToPongError(Message message) {
+    return new MessageOut(
+        message.getTransactionId().toString(),
+        new PayloadError("ping", message.isError()));
   }
 }
